requirements.txt
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3

py~=1.11.0
colorama~=0.4.4
cryptography~=37.0.2
pip~=22.1.1
attrs~=21.4.0
toml~=0.10.2
Pillow~=8.3.1
tornado~=6.1
Jinja2~=3.1.2
pytz~=2022.1
sqlparse~=0.4.2
pluggy~=0.13.1
iniconfig~=1.1.1
atomicwrites~=1.4.0
certifi~=2021.10.8
six~=1.16.0
idna~=3.3
urllib3~=1.26.9
pyparsing~=3.0.9
setuptools~=47.1.0
coreapi~=2.3.3
coreschema~=0.0.4
django-filter~=21.1
uritemplate~=4.1.1
zipp~=3.8.0